+++ src/libs/antares/study/CMakeLists.txt
@@ -73,6 +73,8 @@ set(SRC_STUDY_PART_THERMAL
         include/antares/study/parts/thermal/cluster.h
         include/antares/study/parts/thermal/cluster.hxx
         parts/thermal/cluster.cpp
+	parts/thermal/scenarized_cost_provider.cpp
+	parts/thermal/constant_cost_provider.cpp
         include/antares/study/parts/thermal/cluster_list.h
         parts/thermal/cluster_list.cpp
         include/antares/study/parts/thermal/pollutant.h
